pyproject.toml
[project]
name = "langchain-uv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.6",
    "langchain-core>=0.3.28",
    "langchain>=0.3.13",
    "langchain-community>=0.3.13",
    "google-generativeai>=0.8.3",
    "langchain-google-genai>=2.0.7",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
    "langgraph>=0.2.60",
    "faiss-cpu>=1.9.0.post1",
    "sqlmodel>=0.0.22",
    "streamlit>=1.41.1",
    "psycopg2-binary>=2.9.10",
    "pymongo>=4.9.2",
    "langgraph-checkpoint>=2.0.9",
    "langgraph-checkpoint-mongodb>=0.1.0",
    "crewai>=0.100.1",
    "crewai-tools>=0.33.0",
]
